Apply TopAppBar do not composition in DetailScreen
