+++ app/src/main/java/com/jik/movie/navigation/TopLevelDestination.kt
@@ -1,0 +1,9 @@
+package com.jik.movie.navigation
+
+import com.jik.feature.popular.navigation.PopularNavigation
+
+enum class TopLevelDestination(
+    val route: String
+) {
+    POPULAR(PopularNavigation.route)
+}
