+++ core-ui/src/main/res/values/string.xml
@@ -1,3 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
 <resources>
+
     <string name="app_name">Movie</string>
 </resources>
